requirements.txt
pyenchant
pytest